Langchain_prompts 1/requirements.txt
#langchain core 
langchain
langchain-core

#openAI integration 
langchain-openai
openai

#Anthropic integration
langchain-anthropic


#google gemini (PaLM) integration
langchain-google-genai
google-generativeai

#hugging face integration
langchain-huggingface
transformers
huggingface-hub

# Environment variables management
python-dotenv

#machine learning and data processing
scikit-learn
numpy

streamlit